update README.md: Explain example before showing its output
Previously we had the example output first, and then an explanatory
blurb. Because the introductory blurb mentions the 'at -l' command, having the
example blurb come after the example output made it difficult to determine
whether we were looking at the output of 'at -l' or the output of 'atat' until
we read further.