add file: NOTES.org
I've been doing some research around the lineage of the at(1) command on
GNU/Linux, which has lead to some related history on NetBSD, FreeBSD, OpenBSD,
and Dragonfly BSD. While it started out as an innocent question of "where is
the authoritative upstream to which I should send any patches", it actually
turned into a bit of a digital archaeological dig.

I was considering implementing support for the 'TIME_STYLE' environment
variable that is honored by GNU ls(1), GNU du(1), and possibly other
tools. With such support directly in du(1), there would be no need for atat(1)
or other work-alikes. This is just at the idea stage, and I might decide
against it, but for now I'm going to start collecting my thoughts about it and
relevant tidbits from my research in the NOTES.org file.