atat.in: add first cut at man page (in pod)
Note that we mention SUSv4 and POSIX, but only by way of referencing where the
default of 'at -l' is prescribed. We therefore put the info about the standard
in a section called 'STANDARDS', as opposed to a section called 'CONFORMING
TO' (which is what is generally recommended by man-pages(7)).

Rationale: The atat(1) program is deliberately not conforming to the standard,
but the input on which it operates is predicatable because it /is/ presribed
by the standard.