atat.in: pass-through unrecognized opts to atq(1)
We'll pass through all command line options that are not recognized by 'atat'
(as opposed to exiting with an error).

Same goes for any command line arguments beyond the '--' pseudo-option.

At the moment, this is only useful (I think) for the at(1) options '-V'
(prints the version number) and '-q QUEUE_NAME' (limits output to just that in
the specified job queue).