NOTES.org: org-mode markup tweaks for all "empty" verbatim lines
GitHub renders verbatim lines with content in a way that offsets the text as
you would expect, but GitHub does not recognize verbatim lines that do not
have any content in them as such. However, it /does/ recognize them if there
is a space character after the ':', so that's what we're doing here. Hold your
nose and ignore the trailing whitespace.